Extract getAccountBalanceWithFallback
